Changed MaskValueSet to an interface and added class MaskValueSetImpl
Changed MaskValueSet to an interface and added class MaskValueSetImpl as implementor
